0.1.3+rev219

+++ src/main.cc
@@ -17,7 +17,7 @@ void cxx_main() {
     uart::cout << " * INITRD located at 0x" << bootboot.initrd_ptr << '\n';
     uart::cout << " * INITRD size: 0x" << bootboot.initrd_size << '\n';
     uart::cout << " * memory map located at " << (u64)&bootboot.mmap.ptr << '\n';
-    uart::cout << " * ACPI ID (must be zero): " << (u8)acpi::id() << '\n';
+    uart::cout << " * ACPI ID (must be zero): " << (u64)acpi::id() << '\n';
     uart::cout << " * ACPI CPU count: " << (u8)ACPI_CpuCount << '\n';
 
     initrd::init(bootboot);
